🚚 Rename Redux store (settings-tab → settings)
