+++ modules/ppcp-settings/resources/js/data/settings/constants.js
