+++ modules/ppcp-settings/resources/js/data/settings/controls.js
