+++ modules/ppcp-settings/resources/js/data/settings/hooks.js
