+++ modules/ppcp-settings/resources/js/data/settings/index.js
