+++ modules/ppcp-settings/resources/js/data/settings/reducer.js
