+++ modules/ppcp-settings/resources/js/data/settings/resolvers.js
