+++ modules/ppcp-settings/resources/js/data/settings/selectors.js
